requirements.txt
# Core dependencies
streamlit>=1.28.0
pandas>=2.0.0
numpy>=1.24.0
requests>=2.31.0
python-dateutil>=2.8.2

# Data visualization
matplotlib>=3.7.0
scipy>=1.11.0

# Database
psycopg2-binary>=2.9.9
timescaledb>=0.1.0

# Timezone handling
pytz>=2023.3

# Configuration
toml>=0.10.2
python-dotenv>=1.0.0

# Auto-refresh for real-time updates
streamlit-autorefresh>=1.0.1

# WebSocket support
websockets>=12.0

# FastAPI Backend (for Next.js UI)
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
aiofiles>=23.2.1
